Refactor enemy movement
Split the move between points logic in three different scripts

- one to define the direction of the movement
- other to make a linear movement, reading a direction from another site
- and other to store that direction

The Directable component stores the direction and has an action to fire
when the direction changes.

The LinearMovement component moves linearly based on a speed and a
direction from a Directable. This moves afecting the transform position
directly. It can be changed with a new component that moves the
gameobject applying forces to the rigidbody.

The SteerThroughPoints changes the direction of a Directable towards a
point, and when reaches that point changes the direction to the other
point. (Only has two points by now).

The old MoveBetween is removed and replaced with these new scripts in
order to improve modularity and reusability.

The LinearMovement speed needs to be balanced with the
SteerThroughPoints tolerance to kwnow if it reached the next point.

Cleaned Enemy behaviour

Add UpdateSpriteOnDirectionChange to flip the sprite depending on the
direction